Beta Version 0.3 - RoomDB for Register and Login now Working
