--- app/src/main/java/com/example/kamatechsmobileapplication/LoginActivity.kt
@@ -1,54 +1,0 @@
-package com.example.kamatechsmobileapplication
-
-import android.content.Intent
-import android.os.Bundle
-import android.widget.Button
-import android.widget.EditText
-import android.widget.TextView
-import android.widget.Toast
-import androidx.appcompat.app.AppCompatActivity
-import com.google.firebase.auth.FirebaseAuth
-
-class LoginActivity : AppCompatActivity() {
-    private lateinit var tvRedirectSignUp: TextView
-    lateinit var etEmail: EditText
-    private lateinit var etPass: EditText
-    lateinit var btnLogin: Button
-
-    lateinit var auth: FirebaseAuth
-
-    override fun onCreate(savedInstanceState: Bundle?) {
-        super.onCreate(savedInstanceState)
-        setContentView(R.layout.activity_login)
-
-        tvRedirectSignUp = findViewById(R.id.tvRedirectSignUp)
-        btnLogin = findViewById(R.id.btnLogin)
-        etEmail = findViewById(R.id.etEmailAddress)
-        etPass = findViewById(R.id.etPassword)
-
-        auth = FirebaseAuth.getInstance()
-
-        btnLogin.setOnClickListener{
-            login()
-        }
-        tvRedirectSignUp.setOnClickListener {
-            val intent = Intent(this, RegisterActivity::class.java)
-            startActivity(intent)
-            finish()
-        }
-    }
-    private fun login(){
-        val email = etEmail.text.toString()
-        val pass = etPass.text.toString()
-
-        auth.signInWithEmailAndPassword(email, pass).addOnCompleteListener(this) {
-            if (it.isSuccessful) {
-                Toast.makeText(this, "Successfully Logged In", Toast.LENGTH_SHORT).show()
-                val intent = Intent(this, MainActivity::class.java)
-                startActivity(intent)
-            } else {
-                Toast.makeText(this, "Log In failed", Toast.LENGTH_SHORT).show()
-            }
-        }
-    }
-}
