--- app/src/main/java/com/example/kamatechsmobileapplication/RegisterActivity.kt
@@ -1,68 +1,0 @@
-package com.example.kamatechsmobileapplication
-
-import android.content.Intent
-import android.os.Bundle
-import android.widget.Button
-import android.widget.EditText
-import android.widget.TextView
-import android.widget.Toast
-import androidx.appcompat.app.AppCompatActivity
-import com.google.firebase.auth.FirebaseAuth
-import com.google.firebase.auth.ktx.auth
-import com.google.firebase.ktx.Firebase
-
-class RegisterActivity : AppCompatActivity() {
-
-    lateinit var etEmail: EditText
-    lateinit var etConfPass: EditText
-    private lateinit var etPass: EditText
-    private lateinit var btnSignup: Button
-    lateinit var tvRedirectLogin: TextView
-
-    private lateinit var auth: FirebaseAuth
-
-    override fun onCreate(savedInstanceState: Bundle?) {
-        super.onCreate(savedInstanceState)
-        setContentView(R.layout.activity_register)
-
-        etEmail = findViewById(R.id.etSEmailAddress)
-        etConfPass = findViewById(R.id.etSConfPassword)
-        etPass = findViewById(R.id.etSPassword)
-        btnSignup = findViewById(R.id.btnSSigned)
-        tvRedirectLogin = findViewById(R.id.tvRedirectLogin)
-
-        auth =  Firebase.auth
-
-        btnSignup.setOnClickListener{
-            registerUser()
-        }
-
-        tvRedirectLogin.setOnClickListener {
-            val intent = Intent(this, LoginActivity::class.java)
-            startActivity(intent)
-        }
-    }
-    private fun registerUser(){
-        val email = etEmail.text.toString()
-        val pass = etPass.text.toString()
-        val confirmPassword = etConfPass.text.toString()
-
-        if (email.isBlank() || pass.isBlank() || confirmPassword.isBlank()){
-            Toast.makeText(this, "Email and password can't be empty", Toast.LENGTH_SHORT).show()
-            return
-        }
-        if (pass != confirmPassword){
-            Toast.makeText(this, "Password does not match", Toast.LENGTH_SHORT).show()
-            return
-        }
-        auth.createUserWithEmailAndPassword(email, pass).addOnCompleteListener(this) {
-            if (it.isSuccessful){
-                Toast.makeText(this, "Successfully Registered", Toast.LENGTH_SHORT).show()
-                val intent = Intent(this, LoginActivity::class.java)
-                startActivity(intent)
-            } else{
-                Toast.makeText(this, "Register is not successful!", Toast.LENGTH_SHORT).show()
-            }
-        }
-    }
-}
