+++ app/src/main/java/com/example/kamatechsmobileapplication/fragments/AboutFragment.kt
@@ -1,19 +1,23 @@
-package com.example.kamatechsmobileapplication
+package com.example.kamatechsmobileapplication.fragments
 
 import android.os.Bundle
 import androidx.fragment.app.Fragment
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
+import com.example.kamatechsmobileapplication.R
 
-class StorageFragment : Fragment() {
-    
+
+class AboutFragment : Fragment() {
+
 
     override fun onCreateView(
         inflater: LayoutInflater, container: ViewGroup?,
         savedInstanceState: Bundle?
     ): View? {
         // Inflate the layout for this fragment
-        return inflater.inflate(R.layout.fragment_storage, container, false)
+        return inflater.inflate(R.layout.fragment_about, container, false)
     }
+
+
 }
