+++ app/src/main/java/com/example/kamatechsmobileapplication/fragments/FAQFragment.kt
@@ -1,13 +1,14 @@
-package com.example.kamatechsmobileapplication
+package com.example.kamatechsmobileapplication.fragments
 
 import android.os.Bundle
 import androidx.fragment.app.Fragment
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
+import com.example.kamatechsmobileapplication.R
 
 
-class AboutFragment : Fragment() {
+class FAQFragment : Fragment() {
 
 
     override fun onCreateView(
@@ -15,7 +16,7 @@ class AboutFragment : Fragment() {
         savedInstanceState: Bundle?
     ): View? {
         // Inflate the layout for this fragment
-        return inflater.inflate(R.layout.fragment_about, container, false)
+        return inflater.inflate(R.layout.fragment_f_a_q, container, false)
     }
