+++ app/src/main/res/layout/fragment_f_a_q.xml
@@ -2,7 +2,7 @@
 <layout xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:app="http://schemas.android.com/apk/res-auto"
     xmlns:tools="http://schemas.android.com/tools"
-    tools:context=".FAQFragment">
+    tools:context=".fragments.FAQFragment">
 
 
     <androidx.constraintlayout.widget.ConstraintLayout
