+++ app/src/main/res/layout/fragment_storage.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".StorageFragment">
+    tools:context=".fragments.StorageFragment">
 
     <androidx.constraintlayout.widget.ConstraintLayout
         android:id="@+id/StorageConstraint"
