+++ app/src/main/res/navigation/navigation.xml
@@ -27,7 +27,7 @@
     </fragment>
     <fragment
         android:id="@+id/homeFragment"
-        android:name="com.example.kamatechsmobileapplication.HomeFragment"
+        android:name="com.example.kamatechsmobileapplication.fragments.HomeFragment"
         android:label="fragment_home"
         tools:layout="@layout/fragment_home" >
         <action
@@ -53,12 +53,12 @@
         tools:layout="@layout/activity_weather" />
     <fragment
         android:id="@+id/aboutFragment"
-        android:name="com.example.kamatechsmobileapplication.AboutFragment"
+        android:name="com.example.kamatechsmobileapplication.fragments.AboutFragment"
         android:label="fragment_about"
         tools:layout="@layout/fragment_about" />
     <fragment
         android:id="@+id/FAQFragment"
-        android:name="com.example.kamatechsmobileapplication.FAQFragment"
+        android:name="com.example.kamatechsmobileapplication.fragments.FAQFragment"
         android:label="fragment_f_a_q"
         tools:layout="@layout/fragment_f_a_q" />
 </navigation>
